backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
PyJWT>=2.8.0,<3.0.0
